# Conflicts:
#	Application-Code/app/src/main/java/com/example/medication_reminder_android_app/MainActivity.java

+++ Application-Code/app/src/main/java/com/example/medication_reminder_android_app/UserInputHandler/InputHandler.java
@@ -18,7 +18,7 @@ public abstract class InputHandler {
 
     }
 
-    abstract int inputRequest(Map<String,String> info);
+    abstract long inputRequest(Map<String,String> info);
 
     abstract void deleteRequest(String name);
